Fixed some more shit

+++ server/src/Util/PrefManager.cpp
@@ -241,7 +241,7 @@ String PrefManager::getStringValue(const String&name) const
 ArrayList<String> PrefManager::getKeys()
 {
 	ArrayList<String> keys;
-	for (int i = 0; i < prefs.size(); i++)
+	for (size_t i = 0; i < prefs.size(); i++)
 	{
 		keys.add(prefs.get(i).name);
 	}
